auth/requirements.txt
fastapi
uvicorn
# python-dotenv# if you use .env
# pydantic# optional, often comes with FastAPIannotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astroid==4.0.4
click==8.3.1
dill==0.4.1
exceptiongroup==1.3.1
fastapi==0.128.0
h11==0.16.0
idna==3.11
isort==7.0.0
mccabe==0.7.0
platformdirs==4.5.1
pydantic==2.12.5
pydantic_core==2.41.5
pylint==4.0.4
starlette==0.50.0
tomli==2.4.0
tomlkit==0.14.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astroid==4.0.4
bcrypt==5.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
dill==0.4.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
Faker==40.8.0
fastapi==0.128.0
greenlet==3.3.2
h11==0.16.0
idna==3.11
isort==7.0.0
jose==1.0.0
lance-namespace==0.4.5
lance-namespace-urllib3-client==0.4.5
loguru==0.7.3
Mako==1.3.10
MarkupSafe==3.0.3
mccabe==0.7.0
numpy==2.2.6
passlib==1.7.4
platformdirs==4.5.1
psycopg2-binary==2.9.11
pyarrow==23.0.0
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.12.1
pylance==2.0.0
pylint==4.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
requests==2.32.5
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.48
starlette==0.50.0
tomli==2.4.0
tomlkit==0.14.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astroid==4.0.4
bcrypt==5.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
dill==0.4.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
Faker==40.8.0
fastapi==0.128.0
greenlet==3.3.2
h11==0.16.0
idna==3.11
isort==7.0.0
jose==1.0.0
lance-namespace==0.4.5
lance-namespace-urllib3-client==0.4.5
loguru==0.7.3
Mako==1.3.10
MarkupSafe==3.0.3
mccabe==0.7.0
numpy==2.2.6
passlib==1.7.4
platformdirs==4.5.1
psycopg2-binary==2.9.11
pyarrow==23.0.0
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.12.1
pylance==2.0.0
pylint==4.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
requests==2.32.5
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.48
starlette==0.50.0
tomli==2.4.0
tomlkit==0.14.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astroid==4.0.4
bcrypt==5.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
dill==0.4.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
Faker==40.8.0
fastapi==0.128.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
isort==7.0.0
jose==1.0.0
lance-namespace==0.4.5
lance-namespace-urllib3-client==0.4.5
loguru==0.7.3
Mako==1.3.10
MarkupSafe==3.0.3
mccabe==0.7.0
numpy==2.2.6
passlib==1.7.4
platformdirs==4.5.1
psycopg2-binary==2.9.11
pyarrow==23.0.0
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.12.1
pylance==2.0.0
pylint==4.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
requests==2.32.5
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.48
starlette==0.50.0
tomli==2.4.0
tomlkit==0.14.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astroid==4.0.4
bcrypt==5.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
dill==0.4.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
Faker==40.8.0
fastapi==0.128.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
isort==7.0.0
jose==1.0.0
lance-namespace==0.4.5
lance-namespace-urllib3-client==0.4.5
loguru==0.7.3
Mako==1.3.10
MarkupSafe==3.0.3
mccabe==0.7.0
numpy==2.2.6
passlib==1.7.4
platformdirs==4.5.1
psycopg2-binary==2.9.11
pyarrow==23.0.0
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.12.1
pylance==2.0.0
pylint==4.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
requests==2.32.5
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.48
starlette==0.50.0
tomli==2.4.0
tomlkit==0.14.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
